pip_requirements.txt
Django>=1.4
PIL==1.1.7
South==0.7.6
django-piston==0.2.3
ipython==0.13.1
readline==6.2.4.1
wsgiref==0.1.2
django-extensions==1.0.1
Werkzeug==0.8.3
